Remove Govuk-Use-Recommended-Related-Links header
As an outcome of RFC-163, we are trying to simplify our CDN configuration by
removing functionality that’s no longer used, or by moving logic to other parts
of the stack.

This change was suggested in a comment on RFC-163
(https://github.com/alphagov/govuk-rfcs/pull/163#discussion_r1289816022):

"I think the flag for Whitehall recommended links should be removed. It was put
in as a safety mechanism for the introduction of the links, in case something
really inappropriate was found. There are ways to manually override the links
now, and we've never used the feature flag."

The application logic dependent on this header is removed in
https://github.com/alphagov/government-frontend/pull/2918

It's now recommended to create feature flags as environment variables in helm
charts.