Update the % for ElasticSearch 6.7 AA test
The AA test is currently deployed at a 50/50 variant split. This PR updates the AA test to throttle the split at 0.5%.
The rendering app code has been updated in this PR to handle the C variant[1].

The C variant has been added to make the difference between the CDN traffic allocation and anything that falls into the Z variant (handled by the rendering app) more explicit.

This PR was originally opened on the govuk-cdn-config repo which has been archived. Therefore it has been re-added here and I will close the other PR.
[1]https://github.com/alphagov/finder-frontend/pull/3149