Set the host for DGU proxied services
"Fastly uses a cache key of 'req.url, req.http.host' when deciding which
key to store a cache under. So we should set the host when proxying to
services other than the main data.gov.uk service so that they don't get
cached over data.gov.uk. This happens because we rewrite the URLs to
remove the prefix part of the URL for the proxy service. Setting the
host doesn't change the host of the backend origin service we're
connecting to, it just changes the value of the header sent to the
origin service."

Migrated from:
https://github.com/alphagov/govuk-aws/blob/3f7f92c/terraform/projects/fastly-datagovuk/main.tf#L67-L83

Commits that introduced it:
https://github.com/alphagov/govuk-aws/commit/9f93f5b0499fafa8a410f63417fda704a0ef37f7
https://github.com/alphagov/govuk-aws/commit/3cad66761e3a7e282b04e6770ac9dfc0aa2e7f89

This could be moved to a dynamic block.