Get the Fastly CDN logs for DGU into Splunk
So that the Cyber Defence team can actively monitor in the future
and/or better investigate incidents.

Configured for production and integration, which is the same as www.

We are using an existing index, which has the data retention configured
as 120 days.

Data comes from: https://github.com/alphagov/govuk-fastly-secrets/commit/5055cb0a899e24a5bbafdc1b9422734fd99a6bb4