Turn off VertexSearch A/B test
Temporarily turns off the VertexSearch A/B test.

We've found through testing that some important documents are being
removed from the Google index, even though they are being updated in
the datastore correctly.

Until this problem is fixed it's not safe to continue with the A/B test
as some of the documents affected are for self-assessment and the
self-assessment deadline is imminent.