Reject HTTP requests using a non-standard method
At the moment, request with methods such as DEBUG or PROPFIND will be
passed through the CDN and hit the origin, where they'll be rejected by
nginx as 501s.

This means we end up with a bunch of 5XX responses in our metrics and
logs, even though there's nothing wrong with our system.

    $ curl -w '\n%{http_code}\n' -X DEBUG https://www.gov.uk
    DEBUG method is not supported
    501

This commit rejects these at the CDN, and uses the 405 HTTP status
instead of 501 to avoid polluting the CDN metrics with unactionable
5XXs.

I've included all standard HTTP methods, and additionally FASTLYPURGE
which is how the non-standard PURGE method appears in VCL[1].

I've also checked which non-standard HTTP methods have appeared in
requests so far in May:

    select method, count(*)
    from fastly_logs.govuk_www
    where year=2024 and month=5
    and method not in ('GET','HEAD','POST','PUT','DELETE','CONNECT','OPTIONS','TRACE','PATCH','FASTLYPURGE')
    group by method
    order by count(*) desc

    #	method	_col1
    1	DEBUG	23086
    2	TENB	14458
    3	PROPFIND	7700
    4	HEADX	4029
    5	TRACK	3
    6	BVRPGMCC	1
    7	CET	1
    8	get	1
    9	INDEX	1

... which is enough evidence for me that this won't block anything
unintentionally.

Temporarily, I've restricted this to integration so I can test it,
instead of just yeeting it into production.

[1]: https://www.fastly.com/documentation/reference/vcl/variables/client-request/req-method/