Add `SearchAutocomplete` AB test
We will shortly be testing a new autocomplete component for site search.

This sets up the AB test with an initial 0/0/100 split (all baseline/
"Z"), to allow us to get the rest of the AB test plumbing into place and
use it as a mini feature flag to start testing the functionality on
production.