Return 503 on chat homepage when chat is disabled
This rule is in place to return a 503 if the `disable_chat` variable is
`true` and the request URL is for any GOV.UK Chat page.

However due to the trailing slash, it doesn't apply on the chat homepage
(www.gov.uk/chat).

This updates the logic so we inspect just the path and will kick in if
the URL is "/chat" or "/chat/*".