Send 1% of traffic to ministers index A/B test variant
This will add the header for the ministers index A/B variant to 1% of
all overall traffic, therefore meaning the ministers index page will
start receiving a very low amount of traffic.

We will monitor this and increase the traffic level when we are happy
the page is performing correctly.