Activate observability features on gov.uk services
These toggle on the [observability features](https://www.fastly.com/documentation/guides/observability/) we've paid for.

As we paid for a quota on these features, I'm purposely only activating on the services I'm especially interested in at the moment.

Bouncer has a _lot_ of domains, so it's probably worth a conversation with Fastly before we activate Domain Inspector on that service.