Hardcoded redirect for /prepare
This is a hardcoded 307 redirect that is served direct from the CDN. It will be cached by browsers for 60s to allow us a relatively quick changeover if required.

I've tested this in [Fastly Fiddle](https://fiddle.fastly.dev/fiddle/b11d2c77), and it seems OK. I used the [Fastly docs for redirects](https://www.fastly.com/documentation/solutions/tutorials/redirects/) as reference.

We've been using error codes in the [800 range, which are actually reserved for Fastly](https://www.fastly.com/documentation/reference/vcl/statements/error/#best-practices-for-using-status-codes-for-errors), so I've started this in the [allowed 600 range](https://www.fastly.com/documentation/reference/http/http-statuses/#600-699-available-for-customer-use) to be safe.

I chose 607 as the custom error code because I chose 307 as the redirect code. It's [a temporary redirect](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status/307). Most requests will likely be GET, but some might not be.

The [`synthetic` statement](https://www.fastly.com/documentation/reference/vcl/statements/synthetic/) sets the response body to an empty string, which is what we want.

If we want to retain this capability, we should use a dictionary and generalise it. But I don't intend to do that now.