Align on commenting subroutines
I noticed that the location of comments for subroutines was fairly inconsistent.

A comment at the top of each subroutine appears to "stick" to the top of VSCode, which is handy. It was not the case for one buried halfway in.

I'm starting by working on the www service to limit the load on my brain, but also because the www is the most complicated, so will get most benefit from better hygiene.