Implement rate limiting logic with an IP allow list in VCL
We have rate limiting enabled and configured through the web UI, but that
doesn't allow us to do IP allow listing. Instead, we do it in VCL.

This should have no impact for as long as the
"ratelimit_allow_list_ip_addresses" secret is not set, which is done out of
band from this repo.